backend/requirements.txt
asgiref==3.8.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==5.0.6
djangorestframework==3.15.2
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
setuptools==68.0.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.0.4
wheel==0.40.0
